refactor: move service fn to repo
